requirements.txt
FAdo==1.3.5.1
lark-parser==0.11.3
regex==2021.7.6